feat: add WorkTime creation page and related components
- Implemented WorkTimeNewPage for creating new work time entries.
- Added CompanyFormSections component for managing company details.
- Introduced ConfirmDialog component for confirmation prompts.
- Created EntityDocumentsTab for handling document uploads and management.
- Added Toast component for displaying success and error messages.

+++ backend/src/main/java/com/wagepayroll/liquibase/task/DdlM8WorkTime1.java
@@ -30,7 +30,7 @@ public class DdlM8WorkTime1 implements CustomTaskChange {
 						  name VARCHAR(120) NOT NULL,
 						  code VARCHAR(40) NOT NULL,
 						  hours_per_day DECIMAL(4,2) NOT NULL,
-						  work_days_per_week SMALLINT NOT NULL,
+						  work_days_per_week INT NOT NULL,
 						  description VARCHAR(500) NULL,
 						  active BOOLEAN NOT NULL,
 						  created_at TIMESTAMP NOT NULL,
